+++ backend/src/main/java/backend/mulkkam/auth/dto/KakaoSigninRequest.java
@@ -1,6 +1,6 @@
 package backend.mulkkam.auth.dto;
 
 public record KakaoSigninRequest(
-        String accessToken
+        String oauthAccessToken
 ) {
 }
